// src/shared/manage-errors/code-erros/error-codes.enum.ts
export enum ErrorCodesEnum {
  MON000 = 'MON000',
  MON001 = 'MON001',
  MON002 = 'MON002',
  MON003 = 'MON003',
  MON004 = 'MON004',
  MON005 = 'MON005',
  MON006 = 'MON006',
  MON007 = 'MON007',
  MON008 = 'MON008',
  MON009 = 'MON009',
  MON010 = 'MON010',
  MON011 = 'MON011',
  MON012 = 'MON012',
  MON013 = 'MON013',
  MON014 = 'MON014',
  MON015 = 'MON015',
  MON016 = 'MON016',
  MON017 = 'MON017',
  MON018 = 'MON018',
  MON019 = 'MON019',
  MON020 = 'MON020',
  MON021 = 'MON021',
  MON022 = 'MON022',
  MON023 = 'MON023',
  MON024 = 'MON024',
  MON025 = 'MON025',
  MON026 = 'MON026',
  MON027 = 'MON027',
  MON028 = 'MON028',
  MON029 = 'MON029',
  MON030 = 'MON030',
}